% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddInterval.R
\name{AddInterval}
\alias{AddInterval}
\title{AddInterval}
\usage{
AddInterval(intervals, interval, idx = NULL, minYear = 1980L, maxYear = 2016L)
}
\arguments{
\item{intervals}{intervals}

\item{interval}{interval}

\item{idx}{idx}

\item{minYear}{minYear}

\item{maxYear}{maxYear}
}
\value{
data.table
}
\description{
AddInterval
}
