% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddQuoteMarks.R
\name{AddQuoteMarks}
\alias{AddQuoteMarks}
\title{AddQuoteMarks}
\usage{
AddQuoteMarks(text)
}
\arguments{
\item{text}{Text to be quoted. Required.}
}
\value{
string
}
\description{
Adds quote marks around the passed string.
}
\examples{
AddQuoteMarks("text")

}
