% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyOriginGroupingMap.R
\name{ApplyOriginGroupingMap}
\alias{ApplyOriginGroupingMap}
\title{ApplyOriginGroupingMap}
\usage{
ApplyOriginGroupingMap(inputData, map)
}
\arguments{
\item{inputData}{Input data. Required.}

\item{map}{Data.table object with mapping from RegionOfOrigin to
GroupOfOrigin. Required.}
}
\value{
inputData
}
\description{
Applies RegionOfOrigin grouping map to the input data
}
\examples{
\dontrun{
ApplyOriginGroupingMap(inputData, map)
}

}
