% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToOrdinalNumber.R
\name{ConvertToOrdinalNumber}
\alias{ConvertToOrdinalNumber}
\title{ConvertToOrdinalNumber}
\usage{
ConvertToOrdinalNumber(cardinalNum)
}
\arguments{
\item{cardinalNum}{Integer vector. Required.}
}
\value{
Character vector.
}
\description{
Convert from cardinal to ordinal number.
}
\examples{
ConvertToOrdinalNumber(cardinalNum = c(1, 2, 3, 11))

}
