% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsurePackages.R
\name{EnsurePackages}
\alias{EnsurePackages}
\title{EnsurePackages}
\usage{
EnsurePackages(packageNames)
}
\arguments{
\item{packageNames}{Vector of package names to load. Required.}
}
\value{
NULL (invisibly)
}
\description{
Load packages. They will be installed if needed.
}
\examples{
\dontrun{
EnsurePackages(packageNames = c("jomo"))
}

}
