% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatDiffTime.R
\name{FormatDiffTime}
\alias{FormatDiffTime}
\title{FormatDiffTime}
\usage{
FormatDiffTime(x)
}
\arguments{
\item{x}{\code{\link{difftime}} object. Required.}
}
\value{
string
}
\description{
Pretty printing of time difference in format HH:MM:SS
}
\examples{
startTime <- Sys.time()
endTime <- startTime + 100
FormatDiffTime(x = endTime - startTime)

}
