% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatObjectSize.R
\name{FormatObjectSize}
\alias{FormatObjectSize}
\title{FormatObjectSize}
\usage{
FormatObjectSize(size)
}
\arguments{
\item{size}{Argument name. Required.}
}
\value{
string
}
\description{
Formats object size to human readable format.
}
\examples{
FormatObjectSize(size = 100)
FormatObjectSize(size = 50000)
FormatObjectSize(size = 4000000)

}
