% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatTime.R
\name{FormatTime}
\alias{FormatTime}
\title{FormatTime}
\usage{
FormatTime(x)
}
\arguments{
\item{x}{\code{\link{POSIXct}} object. Required.}
}
\value{
string
}
\description{
Pretty printing of time in format YYYY/MM/DD HH:MM:SS.
}
\examples{
x <- Sys.time()
FormatTime(x)

}
