% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataSummaryArtifacts.R
\name{GetDataSummaryArtifacts}
\alias{GetDataSummaryArtifacts}
\title{GetDataSummaryArtifacts}
\usage{
GetDataSummaryArtifacts(inputData)
}
\arguments{
\item{inputData}{Input data as data.table. Required}
}
\value{
list of objects
}
\description{
Produces plots and artifacts for the data summary page
}
\examples{
\dontrun{
GetDataSummaryArtifacts(inputData)
}

}
