% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDate.R
\name{GetDate}
\alias{GetDate}
\title{y <- 1977
q <- 2
m <- 5
w <- NA
d <- 20
GetDate(y, q, m, w, d)}
\usage{
GetDate(y = NA, q = NA, m = NA, w = NA, d = NA)
}
\description{
y <- 1977
q <- 2
m <- 5
w <- NA
d <- 20
GetDate(y, q, m, w, d)
}
