% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIsState.R
\name{GetIsState}
\alias{GetIsState}
\title{GetIsState}
\usage{
GetIsState(content)
}
\arguments{
\item{content}{object. Required.}
}
\value{
logical
}
\description{
Indicates that the input file is a state file
}
\examples{
\dontrun{
GetIsState(content = "A")
}

}
