% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNotificationQuarterDensityPlot.R
\name{GetNotificationQuarterDensityPlot}
\alias{GetNotificationQuarterDensityPlot}
\title{GetNotificationQuarterDensityPlot}
\usage{
GetNotificationQuarterDensityPlot(plotData, colorPalette = c("#69b023",
  "#7bbcc0", "#9d8b56", "#ce80ce"), xLimits = c(1980, 2025),
  markerLocations = NULL)
}
\arguments{
\item{plotData}{Data table object. Required.}

\item{colorPalette}{Character vector of colors for plotted
series. Optional. Default = \code{c("#c7c7c7", "#69b023", "#7bbcc0",
"#9d8b56", "#ce80ce")}.}

\item{xLimits}{Vector of two numbers (years), lower and upper limit for the
x axis. Optional. Default = \code{c(1980, 2025)}.}

\item{markerLocations}{Vector of numbers (years) for positions of vertical
dashed lines. Optional. Default = \code{NULL}.}
}
\value{
ggplot2 object
}
\description{
Get density plot of notification quarter
}
\examples{
\dontrun{
GetNotificationQuarterDensityPlot(plotData)
}

}
