% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginGroupingMap.R
\name{GetOriginGroupingMap}
\alias{GetOriginGroupingMap}
\title{GetOriginGroupingMap}
\usage{
GetOriginGroupingMap(type, distr)
}
\arguments{
\item{type}{Grouping type}

\item{distr}{Distribution of RegionOfOrigin}
}
\description{
Get mapping from RegionOfOrigin to GroupOfOrigin
}
\examples{
distr <- data.table::data.table(FullRegionOfOrigin = c("REPCOUNTRY", "SUBAFR"), Count = c(1536, 2237))
GetOriginGroupingMap(
  type = "REPCOUNTRY + UNK + 4 most prevalent other regions",
  distr = distr
)

}
