% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPreliminaryAttributesMapping.R
\name{GetPreliminaryAttributesMapping}
\alias{GetPreliminaryAttributesMapping}
\title{GetPreliminaryAttributesMapping}
\usage{
GetPreliminaryAttributesMapping(origData)
}
\arguments{
\item{origData}{Original data. Required.}
}
\value{
list with attributes mapping
}
\description{
Gets attributes mappings.
}
\examples{
\dontrun{
GetPreliminaryAttributesMapping(origData)
}

}
