% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportDefaultParams.R
\name{GetReportDefaultParams}
\alias{GetReportDefaultParams}
\title{GetReportDefaultParams}
\usage{
GetReportDefaultParams(params, skipParamNames = NULL)
}
\arguments{
\item{params}{List of report parameter specifications}

\item{skipParamNames}{Names of parameters to filter out. Optional.
Default = \code{NULL}}
}
\value{
Named list of values
}
\description{
Gets default values of parameters defined in a report file
}
\examples{
\dontrun{
GetReportDefaultParams(paramSpecs, skipParamNames)
}

}
