% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivEstimatesAccuracy.R
\docType{package}
\name{hivEstimatesAccuracy}
\alias{hivEstimatesAccuracy}
\alias{hivEstimatesAccuracy-package}
\title{Improving Accuracy of HIV Estimates in EU/EEA Countries}
\description{
Improves accuracy of HIV estimates in EU/EEA countries.
}
\author{
Creator: Daniel Lewandowski \email{<daniel@nextpagesoft.net>}\cr
Author: Nikos Pantazis \email{<npantaz@med.uoa.gr>}\cr
Author: Magdalena Rosińska \email{<mrosinska@pzh.gov.pl>}\cr
Author: Janusz Janiec \email{<jjaniec@pzh.gov.pl>}\cr
}
