% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeNames.R
\name{MakeNames}
\alias{MakeNames}
\title{MakeNames}
\usage{
MakeNames(strings)
}
\arguments{
\item{strings}{Vector of strings. Required.}
}
\value{
(Named) vector of unique, valid names.
}
\description{
Create a unique vector of names based on strings.
}
\examples{
strings <- c(adjustment1 = "Adjustment 1", adjustment2  = "Fake adjustment")
MakeNames(strings)

strings <- c("Adjustment 1", "Fake adjustment")
MakeNames(strings)

}
