% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivEstimatesAccuracy-package.R
\name{mice.impute.logreg}
\alias{mice.impute.logreg}
\title{mice.impute.logreg}
\usage{
mice.impute.logreg(y, ry, x, wy = NULL, ...)
}
\arguments{
\item{y}{y}

\item{ry}{ry}

\item{x}{x}

\item{wy}{wy}

\item{...}{extra parameters}
}
\description{
mice.impute.logreg
}
