% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivEstimatesAccuracy-package.R
\name{mice.impute.pmm}
\alias{mice.impute.pmm}
\title{mice.impute.pmm}
\usage{
mice.impute.pmm(y, ry, x, wy = NULL, donors = 5L, matchtype = 1L,
  ridge = 1e-05, ...)
}
\arguments{
\item{y}{y}

\item{ry}{ry}

\item{x}{x}

\item{wy}{wy}

\item{donors}{donors}

\item{matchtype}{matchtype}

\item{ridge}{ridge}

\item{...}{extra parameters}
}
\description{
mice.impute.pmm
}
