% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivEstimatesAccuracy-package.R
\name{mice.impute.polyreg}
\alias{mice.impute.polyreg}
\title{mice.impute.polyreg}
\usage{
mice.impute.polyreg(y, ry, x, wy = NULL, nnet.maxit = 100,
  nnet.trace = FALSE, nnet.MaxNWts = 1500, ...)
}
\arguments{
\item{y}{y}

\item{ry}{ry}

\item{x}{x}

\item{wy}{wy}

\item{nnet.maxit}{nnet.maxit}

\item{nnet.trace}{nnet.trace}

\item{nnet.MaxNWts}{nnet.MaxNWts}

\item{...}{extra parameters}
}
\description{
mice.impute.polyreg
}
