% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCMCDiagnostic.R
\name{PlotMCMCDiagnostic}
\alias{PlotMCMCDiagnostic}
\title{PlotMCMCDiagnostic}
\usage{
PlotMCMCDiagnostic(plotData, plotType = "trace", ...)
}
\arguments{
\item{plotData}{Array of matrices. Required.}

\item{plotType}{String denoting type of plot to create. Either "trace" for trace plot, or
"autocorrelation" for autocorrelation plot. Default: "trace".}

\item{...}{Extra parameters passed to \code{\link{plot}} if \code{plotType} is "trace" or to
\code{\link[coda]{autocorr.plot}} if \code{plotType} is "autocorrelation".}
}
\description{
Plot diagnostic information form MCMC chain convergence.
}
\examples{
\dontrun{
PlotMCMCDiagnostic(plotData)
}

}
