% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareDataSetsForModel.R
\name{PrepareDataSetsForModel}
\alias{PrepareDataSetsForModel}
\title{PrepareDataSetsForModel}
\usage{
PrepareDataSetsForModel(dt, splitBy = "Imputation", strata = NULL)
}
\arguments{
\item{dt}{Input data set as data.table object. Required.}

\item{splitBy}{Name of column with values to be used for separation of data sets. Optional.
Default = 'Imputation'}

\item{strata}{Character vector of strata names. Optional. Default = \code{NULL}}
}
\value{
List of HIV models
}
\description{
Prepares data sets for HIV Model
}
\examples{
\dontrun{
  PrepareDataSetsForModel(dt, splitBy = 'Imputation', strata = 'Transmission')
}

}
