% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessInputData.R
\name{PreProcessInputData}
\alias{PreProcessInputData}
\title{PreProcessInputData}
\usage{
PreProcessInputData(inputData)
}
\arguments{
\item{inputData}{Input data. Required.}
}
\value{
data.table object
}
\description{
Pre-processes input data before passing it to adjustment scripts.
}
\examples{
\dontrun{
PreProcessInputData(inputData)
}

}
