% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessInputDataForMice.R
\name{PreProcessInputDataForMice}
\alias{PreProcessInputDataForMice}
\title{PreProcessInputDataForMice}
\usage{
PreProcessInputDataForMice(inputData)
}
\arguments{
\item{inputData}{Input data. Required.}
}
\value{
data.table object
}
\description{
Pre-processes input data before passing it to mice adjustment scripts.
}
\examples{
\dontrun{
PreProcessInputDataForMice(inputData)
}

}
