% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAdjustments.R
\name{RunAdjustments}
\alias{RunAdjustments}
\title{RunAdjustments}
\usage{
RunAdjustments(data, adjustmentSpecs = list())
}
\arguments{
\item{data}{data table object on which adjustments should be applied. Required.}

\item{adjustmentSpecs}{List of adjustment specifications to execute. Optional. Default = \code{list()}.}
}
\value{
data table object after adjustments applied.
}
\description{
Execute adjustments specified in \code{adjustmentFileNames}.
}
\examples{
\dontrun{
RunAdjustments(data)
}

}
