% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRdsFile.R
\name{WriteRdsFile}
\alias{WriteRdsFile}
\title{WriteRdsFile}
\usage{
WriteRdsFile(data, fileName, ...)
}
\arguments{
\item{data}{Object to save. Required.}

\item{fileName}{Name of the saved file. Required.}

\item{...}{Additional parameters passed to \code{saveRDS}. Optional.}
}
\value{
NULL (invisibly)
}
\description{
Write an object to a R rds file.
}
\examples{
\dontrun{
WriteRdsFile(data, fileName)
}

}
