% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginGroupingMap.R
\name{GetOriginGroupingMap}
\alias{GetOriginGroupingMap}
\title{GetOriginGroupingMap}
\usage{
GetOriginGroupingMap(type, distr, groups)
}
\arguments{
\item{type}{Grouping type}

\item{distr}{Distribution of RegionOfOrigin}

\item{groups}{Groups}
}
\description{
Get mapping from RegionOfOrigin to GroupOfOrigin
}
\examples{
distr <- data.table::data.table(
  FullRegionOfOrigin = c("REPCOUNTRY", "SUBAFR"),
  Count = c(1536, 2237))
GetOriginGroupingMap(
  type = "REPCOUNTRY + UNK + 3 most prevalent regions + OTHER",
  distr = distr
)

}
