% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareDataSetsForModel.R
\name{PrepareDataSetsForModel}
\alias{PrepareDataSetsForModel}
\title{PrepareDataSetsForModel}
\usage{
PrepareDataSetsForModel(dt, by = c())
}
\arguments{
\item{dt}{Input data set as data.table object. Required.}

\item{by}{Character vector of strata names. Optional. Def = \code{c()}}
}
\value{
List of HIV models
}
\description{
Prepares data sets for HIV Model
}
\examples{
PrepareDataSetsForModel(dt, by = 'transmission')

}
