% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunApp.R
\name{RunApp}
\alias{RunApp}
\title{RunApp}
\usage{
RunApp(launchBrowser = TRUE)
}
\arguments{
\item{launchBrowser}{Logical indicating to open the app in a newly open web browser}
}
\value{
NULL (invisibly)
}
\description{
Run the application.
}
\examples{
\dontrun{
RunApp()
}

}
