% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunApp.R
\name{RunApp}
\alias{RunApp}
\title{RunApp}
\usage{
RunApp()
}
\value{
NULL (invisibly)
}
\description{
Run the whole application. A Shiny application will start in the default web browser.
}
\examples{
\dontrun{
RunApp()
}

}
